Renames folder from 'ui' to 'shared'
